+++ package.json
@@ -16,10 +16,10 @@
     "story-bot": "file:"
   },
   "_moduleAliases": {
-    "@app": "src/app",
     "@commands": "src/commands",
     "@data": "src/data",
-    "@model": "src/model"
+    "@model": "src/model",
+    "@service": "src/service"
   },
   "devDependencies": {
     "@flydotio/dockerfile": "^0.5.8"
